Refactor to match new model.
We no longer have the concept of approved/not approved. That was a flawed
assumption since different licenses may be OK under different circumstances.

Instead, we will tell the app (using a config parameter) the preferred
service model for the application, and we'll be told if we're compliant or
not.